AVL-based sampling

+++ include/util/record.h
@@ -8,7 +8,7 @@ namespace lsm {
 typedef uint32_t rec_hdr;
 typedef uint64_t key_type;
 typedef uint32_t value_type;
-typedef double   weight_type;
+typedef uint64_t weight_type;
 
 constexpr static size_t key_size = sizeof(key_type);
 constexpr static size_t value_size = sizeof(value_type);
